create HandListener class as super class for AppListener and CollectListener
